chore: automated code quality fixes
- ruff format: 7 files reformatted
- ruff --fix/--unsafe-fixes: 21 auto-fixable issues resolved (I001, UP037, UP042, B905, BLE001, RUF100)
- SIM102: collapsed nested if in victoria_node.py macro bias check
- mypy fixes: wrapped no-any-return in overnight_runner.py, curvature_signal.py; str() for adversarial flags in orchestrator_v2.py; lambda key in training_watchdog.py; removed stale type: ignore in wavelet_signal.py

+++ omega/nodes/victoria/curvature_signal.py
@@ -259,4 +259,4 @@ class GeodesicCurvatureSignal:
         mean = sum(h) / len(h)
         var = sum((x - mean) ** 2 for x in h) / max(1, len(h) - 1)
         std = math.sqrt(var) if var > 0 else _EPS
-        return (value - mean) / std
+        return float((value - mean) / std)
